Update MedicalVolume.to_sitk
The current implementation worked somewhat to fix errors with specific image
acquisitions, but was somewhat specific to sagittal images. It has now been
updated to reformat and then export images to sitk as would be expected
for a specific image acquisition type (e.g., sagittal, coronal, axial).

Essentially, based on if its sag, ax, or coral, the `mv` will be reformatted.
Then, this reformatted image along with the affine/spacing information
a new sitk Image will be created. Finally, if the dicom was not collected in
this standard orientation, the user can still choose to flip the volume along
any of the three axes using boolean flags.

New tests were also written to ensure the reformatting is such that the
image is roughly what it would be like if SITK itself loaded it.